// src/types/api/postLexiangUploadBodyMediaType.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Genkit AI Service API
 * AI 模型提供商管理服务 API 文档
提供模型提供商、模型信息和参数规则的查询接口
 * OpenAPI spec version: 1.0.0
 */

/**
 * 媒体类型
 */
export type PostLexiangUploadBodyMediaType =
  (typeof PostLexiangUploadBodyMediaType)[keyof typeof PostLexiangUploadBodyMediaType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostLexiangUploadBodyMediaType = {
  file: 'file',
  video: 'video',
  audio: 'audio',
} as const;